+++ src/components/Handbook/CostTable.tsx
