+++ src/components/Handbook/Handbook.tsx
@@ -26,14 +26,14 @@ interface InfoProps {
 // * Navigate through both using the buttons provided at the top of the container.
 // ――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――
 
-const Info = ({ symbols, selectedSymbol, swapped }: InfoProps) => {
+const Handbook = ({ symbols, selectedSymbol, swapped }: InfoProps) => {
   const isMobile = useMediaQuery({ query: `(max-width: 767px)` });
 
   const [selectedInfo, setSelectedInfo] = useState(1);
 
   return (
     <section className="flex justify-center mx-4 md:mx-8">
-      <div className="bg-gradient-to-t from-card-tool to-card-grad rounded-lg py-8 md:py-10 w-[360px] md:w-full h-[650px] md:h-[700px] max-w-[800px]">
+      <div className="bg-gradient-to-t from-card-tool to-card-grad rounded-lg py-8 md:py-10 h-[650px] md:h-[700px] w-[360px] md:w-full max-w-[800px]">
         {/* NAVBAR */}
         <nav className="flex text-center bg-dark shadow-input transition-all">
           <SlideButton
@@ -69,4 +69,4 @@ const Info = ({ symbols, selectedSymbol, swapped }: InfoProps) => {
   );
 };
 
-export default Info;
+export default Handbook;
