+++ src/components/Handbook/RatioTable.tsx
