--- src/components/Selector/Selector.css
@@ -1,3 +1,0 @@
-.selector {
-    @apply flex justify-center
-}
